Some fixes
Got ADCS model to work with WBC and PD. Reviewed model to check for joe's concerns on how we were zenith poining (torq parallel to B) and swapped out out geomag model. Also output of control block should be a bus, more info is needed in actuator block. added a bus on the input to the satellite block (so that total and act torq can be distinguished). Actuator torque is feed back to control block. needed for WBC.