fixed zenith pointing bug
Fixed a bug in how the B field was converted from inertial to body frame. This what was allowing us to do control when in an equitorial orbit with a constant mag field pointing north-south. Can nolonger get the needed torque in that scenario as expected. Can still do zenith pointing with IGRF model. IT just takes longer since we're relying on the small componetnts of B field in equitorial plane.